Remove "bi-directional transport"
The mention of "bi-directional transport" when the subscription pushes notifications to the client was slightly confusing. Removed the wording since the logical communication flow is obviously bi-directional. This can be done via an inherently bi-directional transport like WebSockets, or it can be composed of two uni-directional transports (pair of HTTP connections). Either way, this will not be specified in the spec language and is considered an implementation detail.