Add Alan to attendees and live queries to agenda
I'll be attending in-person. But I assumed location to be where I work. Not sure!

I didn't want to block too much time for discussion of live queries. I think 10 mins should be sufficient to share my findings and what's left to figure out.

I'm happy to take notes on another section of the agenda.