Add upcoming schedule (#137)
Now that the GraphQL Foundation is formed, I'm bringing my attention back to the Working Group meetings. I know there hasn't been much motion on the spec or working group for the last 6 months while we've been figuring things out. I really appreciate everyone's patience and know we have a lot to cover and catch up on.

I'm proposing here a monthly cadence: the first Thursday of each month. Because our biggest growing community cohort is in europe and the middle or east coast US, I'm also proposing moving the times back to 9am pacific. I understand that this is very much not an ideal time for asia/pacific/austrialia. Frankly, it's near impossible to pick a time that doesn't overlap late evening hours for someone in the community. We may adjust times for some future meetings to balance the pain. I'm open to feedback.

This also updates the readme and guidelines in each agenda to include links to https://github.com/graphql/foundation with references to the legal document you must sign to attend. *Please sign this before the next WG meeting*, I know most regular attendees have done so already.