Add  Object Identity to discussion (#245)
Facebook is starting to create a notion of object identity encoded and enforced in the schema. I can discuss why, but we aren't out of experimenting phase, so aren't yet ready to create a spec RFC. Right now, the concept is more of a "plug in" to the existing spec, that clients like Relay can depend on, rather than requiring any language changes.